Don't use the singleton pattern
The singleton pattern doesn't play well with node.js servers in which multiple
requests may be served to different users in parallel.

For example, if A and B are both users of a site, and are connected via
SoundCloud, it's possible that A and B might be concurrently making requests
that involve the SoundCloud API, and due to the race condition created by the
singleton, B's access token might end up getting used for A's request.